+++ api-lib/frames/personas/PersonaThreeFrame.tsx
@@ -1,33 +1,36 @@
 import React from 'react';
 
-import { OGAvatar } from '../../og/OGAvatar';
+import { Frame } from '../../../_api/frames/router.tsx';
+import { OGAvatar } from '../../../_api/og/OGAvatar.tsx';
 import { getViewerFromParams } from '../_getViewerFromParams.ts';
 import { staticResourceIdentifier } from '../_staticResourceIdentifier.ts';
-import { FrameBgImage, IMAGE_URL_BASE } from '../layoutFragments/FrameBgImage';
-import { FrameBody } from '../layoutFragments/FrameBody';
-import { FrameBodyGradient } from '../layoutFragments/FrameBodyGradient';
-import { FrameFooter } from '../layoutFragments/FrameFooter';
-import { FrameHeadline } from '../layoutFragments/FrameHeadline';
-import { FrameWrapper } from '../layoutFragments/FrameWrapper';
-import { Frame } from '../router';
+import {
+  FrameBgImage,
+  IMAGE_URL_BASE,
+} from '../layoutFragments/FrameBgImage.tsx';
+import { FrameBody } from '../layoutFragments/FrameBody.tsx';
+import { FrameBodyGradient } from '../layoutFragments/FrameBodyGradient.tsx';
+import { FrameFooter } from '../layoutFragments/FrameFooter.tsx';
+import { FrameHeadline } from '../layoutFragments/FrameHeadline.tsx';
+import { FrameWrapper } from '../layoutFragments/FrameWrapper.tsx';
 
 const imageNode = async (params: Record<string, string>) => {
   const { viewerProfile } = await getViewerFromParams(params);
 
   return (
     <FrameWrapper>
-      <FrameBgImage src="persona-2.jpg" />
+      <FrameBgImage src="persona-3.jpg" />
       <FrameBody>
         <FrameBodyGradient
           gradientStyles={{
             background:
-              'radial-gradient(circle at 25% 0%, #0DC0E7 0%, #7908D2 80%)',
+              'radial-gradient(circle at 25% 0%, #3300FF 0%, #EF7200 80%)',
             opacity: 0.7,
           }}
         />
         <FrameHeadline>
           <OGAvatar avatar={viewerProfile?.avatar} />
-          <div tw="flex items-center grow justify-center">Level 2</div>
+          <div tw="flex items-center grow justify-center">Level 3</div>
           <img
             alt="gem"
             src={IMAGE_URL_BASE + 'GemWhite.png'}
@@ -35,23 +38,23 @@ const imageNode = async (params: Record<string, string>) => {
           />
         </FrameHeadline>
         <FrameFooter>
-          Send 5 GIVE
+          Join CoLinks
           <br />
-          to level up more, bruhv
+          to level up again.
         </FrameFooter>
       </FrameBody>
     </FrameWrapper>
   );
 };
 
-export const PersonaTwoFrame: Frame = {
-  id: 'persona2',
+export const PersonaThreeFrame: Frame = {
+  id: 'persona3',
   homeFrame: false,
   imageNode: imageNode,
   resourceIdentifier: staticResourceIdentifier,
   buttons: [
     {
-      title: 'Try @givebot',
+      title: 'Join CoLinks',
       action: 'link',
       target: 'https://colinks.coordinape.com',
     },
